moved modal exports to a group
