+++ src/kirby/modal.d.ts
@@ -1,0 +1,1 @@
+modal.ts
