+++ src/kirby/modal.ts
@@ -1,0 +1,2 @@
+export { ModalController } from './components/modal/services/modal.controller';
+export { ModalConfig } from './components/modal/modal-wrapper/config/modal-config';
